examples/vector_databases/vector-search-db/oracle_db/requirements.txt
oracledb
openai
python-dotenv
numpy
pandas
streamlit